+++ src/config.js
@@ -1,0 +1,3 @@
+const DEBUG = false;
+
+export {DEBUG};